requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arandr==0.1.11
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.1.dev0
autocommand==2.2.2
build==1.2.2
cairocffi==1.6.1
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
dbus-python==1.3.2
dirty-equals==0.8.0
distlib==0.3.9
distro==1.9.0
Django==5.1.5
exceptiongroup==1.2.2
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
greenlet==3.1.1
gscreenshot==3.8.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
inflect==7.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jiter==0.8.2
lensfun==0.3.4
lit==19.1.7.dev0
lsprotocol==2023.0.1
lxml==5.3.0
meson==1.6.1
more-itertools==10.5.0
msgpack==1.0.5
numpy==2.2.2
openai==1.59.8
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pipenv==2024.4.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
psutil==6.1.1
pycairo==1.27.0
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
pygls==1.3.1
Pygments==2.19.1
PyGObject==3.50.0
pynvim==0.5.2
PyOpenGL==3.1.7
pyparsing==3.1.2
pyproject_hooks==1.2.0
PyQt6==6.8.0
PyQt6_sip==13.9.1
PySide6==6.8.1.1
pytest==8.3.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
qtile==0.29.1.dev0+geed1e03c.d20241126
requests==2.32.3
ruff==0.9.2
ruff-lsp==0.0.60
semantic_version==2.10.0
setuptools==75.2.0
setuptools-rust==1.10.2
shiboken6==6.8.1.1
shiboken6-generator==6.8.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
starlette==0.45.2
systemd-python==235
tomli==2.0.1
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
trove-classifiers==2025.1.6.15
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.3.0
userpath==1.9.2
validate-pyproject==0.23.post1.dev0+gf45606b.d20250111
virtualenv==20.28.0
virtualenv-clone==0.5.7
websocket-client==1.8.0
wheel==0.45.0
wsproto==1.2.0
xcffib==1.5.0
